// sdk/node/src/generated/models/OperationArtifactUploadRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * Direct private file declaration. Gregale derives its opaque operation reference and storage key.
 */
export type OperationArtifactUploadRequest = {
  report_id: string;
  name: string;
  size_bytes: number;
  sha256: string;
};